renamed files to match folder name
